+++ src/main/kotlin/ee/carlrobert/codegpt/ui/textarea/PromptTextField.kt
@@ -12,6 +12,7 @@ import com.intellij.openapi.editor.ex.EditorEx
 import com.intellij.openapi.fileTypes.FileTypes
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.TextRange
+import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.wm.ToolWindowManager
 import com.intellij.ui.ComponentUtil.findParentByCondition
 import com.intellij.ui.EditorTextField
@@ -40,7 +41,8 @@ data class AppliedSuggestionActionInlay(
 
 data class AppliedCodeActionInlay(
     override val inlay: Inlay<PromptTextFieldInlayRenderer?>,
-    val code: String
+    val code: String,
+    val editorFile: VirtualFile
 ) : AppliedActionInlay
 
 const val AT_CHAR = '@'
@@ -83,7 +85,7 @@ class PromptTextField(
     fun addInlayElement(
         actionPrefix: String,
         text: String,
-        fileName: String? = null,
+        editorFile: VirtualFile? = null,
         tooltipText: String? = null
     ) {
         editor?.let {
@@ -91,7 +93,7 @@ class PromptTextField(
                 it.caretModel.offset,
                 actionPrefix,
                 text,
-                fileName = fileName,
+                editorFile = editorFile,
                 tooltipText = tooltipText
             )
         }
@@ -102,7 +104,7 @@ class PromptTextField(
         actionPrefix: String,
         text: String?,
         actionItem: SuggestionActionItem? = null,
-        fileName: String? = null,
+        editorFile: VirtualFile? = null,
         tooltipText: String? = null
     ) {
         runUndoTransparentWriteAction {
@@ -115,7 +117,7 @@ class PromptTextField(
                     project,
                     actionPrefix,
                     text,
-                    fileName ?: "",
+                    editorFile?.name ?: "",
                     tooltipText
                 ) { inlay ->
                     appliedInlays.removeIf { appliedInlay -> appliedInlay.inlay == inlay }
@@ -123,10 +125,10 @@ class PromptTextField(
                 })
             if (inlay != null) {
                 // TODO
-                if (tooltipText == null) {
-                    appliedInlays.add(AppliedSuggestionActionInlay(inlay, actionItem))
+                if (tooltipText != null && editorFile != null) {
+                    appliedInlays.add(AppliedCodeActionInlay(inlay, tooltipText, editorFile))
                 } else {
-                    appliedInlays.add(AppliedCodeActionInlay(inlay, tooltipText))
+                    appliedInlays.add(AppliedSuggestionActionInlay(inlay, actionItem))
                 }
                 editor?.caretModel?.moveToOffset(document.textLength)
             }